requirements.txt
# base packages
#python-dateutil
#dds-py
#-e git+https://github.com/tjhunter/dds_py#egg=dds-py
#xlwt

# adls size
# azure-functions==1.12.0
# azure-functions-durable==1.2.2
azure-identity==1.12.0
azure-mgmt-storage==21.0.0
azure-storage-file-datalake==12.9.1

# libraries
diagrams
graphviz
requests
pipenv
#beautifulsoup4
#scikit-learn
#matplotlib
#spacy<3.0.0
#numpy<1.20
#scipy
#jedi<0.18.0
#pandas
#altair

# local libraries
jupyter<=1.0.0

# Testing, typing
flake8
#pytest<=6.2
#mypy==0.790
#black==20.8b1
#data-science-types<0.3


# Testing, typing
pytest
pytest-azurepipelines
pytest-cov
#scikit-learn
#recommonmark
#sphinx_rtd_theme
#databricks-cli
#pandas
#pyspark
#mypy
#flake8
#pyspark-stubs
#jupyter
#pandas
#matplotlib
#black==20.8b1
#seaborn
#streamlit
#dataclasses-json
#ujson
#data-science-types<0.3
